+++ src/media/mod.rs
@@ -16,6 +16,7 @@
 pub mod audio;
 pub mod markers;
 pub mod photo;
+pub mod playlist;
 pub mod streams;
 pub mod tags;
 pub mod video;
@@ -23,6 +24,7 @@ pub mod video;
 pub use audio::{Album, Artist, Track};
 pub use markers::{Chapter, Marker, MarkerKind};
 pub use photo::{Photo, PhotoEntry, Photoalbum};
+pub use playlist::{Playlist, PlaylistKind};
 pub use streams::{
     AudioStream, LyricStream, Media, MediaPart, Stream, StreamCommon, SubtitleStream,
     UnknownStream, VideoStream,